+++ aider/repo.py
@@ -10,7 +10,7 @@ from aider.sendchat import simple_send_with_retries
 
 from .dump import dump  # noqa: F401
 
-ANY_GIT_ERROR = (git.exc.ODBError, git.exc.GitError)
+ANY_GIT_ERROR = (git.exc.ODBError, git.exc.GitError, OSError, IndexError)
 
 
 class GitRepo:
@@ -21,6 +21,7 @@ class GitRepo:
     aider_ignore_last_check = 0
     subtree_only = False
     ignore_file_cache = {}
+    git_repo_error = None
 
     def __init__(
         self,
@@ -257,15 +258,26 @@ class GitRepo:
             commit = self.repo.head.commit
         except ValueError:
             commit = None
+        except ANY_GIT_ERROR as err:
+            self.git_repo_error = err
+            self.io.tool_error(f"Unable to list files in git repo: {err}")
+            self.io.tool_output("Is your git repo corrupted?")
+            return []
 
         files = set()
         if commit:
             if commit in self.tree_files:
                 files = self.tree_files[commit]
             else:
-                for blob in commit.tree.traverse():
-                    if blob.type == "blob":  # blob is a file
-                        files.add(blob.path)
+                try:
+                    for blob in commit.tree.traverse():
+                        if blob.type == "blob":  # blob is a file
+                            files.add(blob.path)
+                except ANY_GIT_ERROR as err:
+                    self.git_repo_error = err
+                    self.io.tool_error(f"Unable to list files in git repo: {err}")
+                    self.io.tool_output("Is your git repo corrupted?")
+                    return []
                 files = set(self.normalize_path(path) for path in files)
                 self.tree_files[commit] = set(files)